+++ tests/test_openapi_paths.py
@@ -4,6 +4,7 @@ from .schemas import FooSchema
 from .schemas import HeaderSchema
 from .schemas import PaginationSchema
 from .schemas import QuerySchema
+from apiflask import doc
 from apiflask import input
 from apiflask import output
 
@@ -234,3 +235,28 @@ def test_register_validation_error_response(app, client):
         error_code] is not None
     assert rv.json['paths']['/bar']['get']['responses'][
         error_code]['description'] == 'Validation error'
+
+
+def test_auto_404_error(app, client):
+    @app.get('/foo')
+    def foo():
+        pass
+
+    @app.get('/bar/<int:id>')
+    def bar():
+        pass
+
+    @app.get('/baz/<int:id>')
+    @doc(responses={404: 'Pet not found'})
+    def baz():
+        pass
+
+    rv = client.get('/openapi.json')
+    assert rv.status_code == 200
+    validate_spec(rv.json)
+    assert '404' not in rv.json['paths']['/foo']['get']['responses']
+    assert '404' in rv.json['paths']['/bar/{id}']['get']['responses']
+    assert rv.json['paths']['/bar/{id}']['get']['responses'][
+        '404']['description'] == 'Not found'
+    assert rv.json['paths']['/baz/{id}']['get']['responses'][
+        '404']['description'] == 'Pet not found'